requirements.txt
absl-py==2.2.2
astroid==3.3.9
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
flake8==7.2.0
flatbuffers==25.2.10
fonttools==4.57.0
idna==3.10
isort==6.0.1
jax==0.6.0
jaxlib==0.6.0
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
mediapipe==0.10.21
ml_dtypes==0.5.1
mypy==1.15.0
mypy_extensions==1.1.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
protobuf==4.25.6
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
pylint==3.3.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
tomlkit==0.13.2
typing_extensions==4.13.2
urllib3==2.4.0
